// frontend/src/company.d.ts
export interface Treatment {
  TreatmentID: number;
  TreatmentName: string;
  Description: string | null;
  Cost: number;
}